go-ship-vn/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.13.0
django-filter==22.1
django-heroku==0.3.1
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-extensions==0.7.1
drf-yasg==1.21.4
gunicorn==20.1.0
heroku==0.1.4
idna==3.4
importlib-metadata==5.0.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
mysqlclient==2.1.1
openapi-codec==1.3.2
packaging==21.3
phonenumbers==8.12.57
Pillow==9.3.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==1.5
python-dotenv==0.19.1
pytz==2022.5
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.17.6
sqlparse==0.4.3
tzdata==2022.5
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
zipp==3.10.0
